fix corner bug, archival
